refactor pure SHA3 API
